feat: remove popup in site map
