--- src/features/SiteMap/MapShow.tsx
@@ -1,413 +1,0 @@
-"use client";
-
-import React, { useState } from "react";
-import Button from "@/components/Button";
-import {
-  getWaterRefillStations,
-  getShowerStations,
-  getRestrooms,
-  getMedicalStations,
-  getAccommodations,
-  type WaterRefillStations,
-  type ShowerStations,
-  type RestRooms,
-  type MedicalStation,
-  type Accommodations,
-} from "@/lib/api";
-import InfoCard from "@/components/InfoCard";
-
-type LocationCategory =
-  | "all"
-  | "water_refill_stations"
-  | "shower_stations"
-  | "restrooms"
-  | "medical_stations"
-  | "accommodations";
-type ShowMode = "mapShow" | "listShow";
-
-const CATEGORIES = [
-  {
-    key: "all",
-    name: "全部",
-  },
-  {
-    key: "water_refill_stations",
-    name: "加水站",
-  },
-  {
-    key: "shower_stations",
-    name: "洗澡點",
-  },
-  {
-    key: "restrooms",
-    name: "廁所",
-  },
-  {
-    key: "medical_stations",
-    name: "醫療站",
-  },
-  {
-    key: "accommodations",
-    name: "住宿",
-  },
-];
-
-const MAP_URL = "https://guangfu250923-map.pttapp.cc/map.html";
-const MAP_HEIGHT = 422;
-
-export default function MapShow() {
-  const [showMode, setShowMode] = useState<ShowMode>("mapShow");
-  const [selectedCategory, setSelectedCategory] =
-    useState<LocationCategory>("all");
-  const [waterRefillStations, setWaterRefillStations] = useState<
-    WaterRefillStations[]
-  >([]);
-  const [showerStations, setShowerStations] = useState<ShowerStations[]>([]);
-  const [restRooms, setRestRooms] = useState<RestRooms[]>([]);
-  const [medicalStations, setMedicalStations] = useState<MedicalStation[]>([]);
-  const [accommodations, setAccommodations] = useState<Accommodations[]>([]);
-  const [allData, setAllData] = useState<
-    (
-      | WaterRefillStations
-      | ShowerStations
-      | RestRooms
-      | MedicalStation
-      | Accommodations
-    )[]
-  >([]);
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  async function fetchWaterRefillStations() {
-    try {
-      setLoading(true);
-      setError(null);
-      const response = await getWaterRefillStations(50, 0);
-      setWaterRefillStations(response.member);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "載入失敗");
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  async function fetchShowerStations() {
-    try {
-      setLoading(true);
-      setError(null);
-      const response = await getShowerStations(50, 0);
-      setShowerStations(response.member);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "載入失敗");
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  async function fetchRestRooms() {
-    try {
-      setLoading(true);
-      setError(null);
-      const response = await getRestrooms(50, 0);
-      setRestRooms(response.member);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "載入失敗");
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  async function fetchMedicalStations() {
-    try {
-      setLoading(true);
-      setError(null);
-      const response = await getMedicalStations(50, 0);
-      setMedicalStations(response.member);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "載入失敗");
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  async function fetchAccommodations() {
-    try {
-      setLoading(true);
-      setError(null);
-      const response = await getAccommodations(50, 0);
-      setAccommodations(response.member);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "載入失敗");
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  async function fetchAll() {
-    try {
-      setLoading(true);
-      setError(null);
-      const [
-        responseWaterRefillStations,
-        responseShowerStations,
-        responseRestrooms,
-        responseMedicalStations,
-        responseAccommodations,
-      ] = await Promise.all([
-        getWaterRefillStations(50, 0),
-        getShowerStations(50, 0),
-        getRestrooms(50, 0),
-        getMedicalStations(50, 0),
-        getAccommodations(50, 0),
-      ]);
-      setWaterRefillStations(responseWaterRefillStations.member);
-      setShowerStations(responseShowerStations.member);
-      setRestRooms(responseRestrooms.member);
-      setMedicalStations(responseMedicalStations.member);
-      setAccommodations(responseAccommodations.member);
-
-      const combined = [
-        ...responseWaterRefillStations.member,
-        ...responseShowerStations.member,
-        ...responseRestrooms.member,
-        ...responseMedicalStations.member,
-        ...responseAccommodations.member,
-      ];
-      combined.sort((a, b) => a.created_at - b.created_at);
-      setAllData(combined);
-    } catch (err) {
-      setError(err instanceof Error ? err.message : "載入失敗");
-    } finally {
-      setLoading(false);
-    }
-  }
-
-  const handleCategoryClick = async (categoryKey: LocationCategory) => {
-    setSelectedCategory(categoryKey);
-    if (categoryKey === "all") {
-      await fetchAll();
-    } else if (categoryKey === "water_refill_stations") {
-      await fetchWaterRefillStations();
-    } else if (categoryKey === "shower_stations") {
-      await fetchShowerStations();
-    } else if (categoryKey === "restrooms") {
-      await fetchRestRooms();
-    } else if (categoryKey === "medical_stations") {
-      await fetchMedicalStations();
-    } else if (categoryKey === "accommodations") {
-      await fetchAccommodations();
-    }
-  };
-
-  const handleModeChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
-    setShowMode(event.target.value as ShowMode);
-    fetchAll();
-  };
-
-  return (
-    <div>
-      <div className="flex my-3">
-        <select
-          className="bg-gray-100 p-3"
-          value={showMode}
-          onChange={handleModeChange}
-        >
-          <option value="mapShow">地圖顯示</option>
-          <option value="listShow">列表顯示</option>
-        </select>
-        {showMode === "listShow" && (
-          <div className="ml-4 flex overflow-y-scroll [scrollbar-width:none]">
-            {CATEGORIES.map((category) => (
-              <Button
-                className="ml-2 border-gray-100"
-                key={category.key}
-                onClick={() =>
-                  handleCategoryClick(category.key as LocationCategory)
-                }
-                active={selectedCategory === category.key}
-              >
-                {category.name}
-              </Button>
-            ))}
-          </div>
-        )}
-      </div>
-      <div>
-        {showMode === "mapShow" && (
-          <iframe
-            src={MAP_URL}
-            title="地圖顯示"
-            width="100%"
-            height={MAP_HEIGHT}
-          />
-        )}
-        {showMode === "listShow" && (
-          <div className="space-y-4">
-            {loading && (
-              <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                載入中...
-              </div>
-            )}
-
-            {error && (
-              <div className="text-center py-8 text-red-500 dark:text-red-400">
-                錯誤: {error}
-              </div>
-            )}
-
-            {!loading && !error && selectedCategory === "all" && (
-              <>
-                {allData.length === 0 ? (
-                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                    此分類暫無資料
-                  </div>
-                ) : (
-                  allData.map((station) => (
-                    <InfoCard
-                      key={station.id}
-                      name={station.name}
-                      address={station.location}
-                      contact={station.phone}
-                      hours={station.opening_hours || ""}
-                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
-                        station.location
-                      )}`}
-                      fullData={station}
-                    />
-                  ))
-                )}
-              </>
-            )}
-
-            {!loading &&
-              !error &&
-              selectedCategory === "water_refill_stations" && (
-                <>
-                  {waterRefillStations.length === 0 ? (
-                    <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                      此分類暫無資料
-                    </div>
-                  ) : (
-                    waterRefillStations.map((station) => (
-                      <InfoCard
-                        key={station.id}
-                        name={station.name}
-                        type={station.water_type}
-                        address={station.location}
-                        contact={station.phone}
-                        hours={station.opening_hours || ""}
-                        mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
-                          station.location
-                        )}`}
-                        fullData={station}
-                      />
-                    ))
-                  )}
-                </>
-              )}
-
-            {!loading && !error && selectedCategory === "shower_stations" && (
-              <>
-                {showerStations.length === 0 ? (
-                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                    此分類暫無資料
-                  </div>
-                ) : (
-                  showerStations.map((station) => (
-                    <InfoCard
-                      key={station.id}
-                      type={station.facility_type}
-                      name={station.name}
-                      address={station.location}
-                      contact={station.phone}
-                      hours={station.time_slots || ""}
-                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
-                        station.location
-                      )}`}
-                      fullData={station}
-                    />
-                  ))
-                )}
-              </>
-            )}
-
-            {!loading && !error && selectedCategory === "restrooms" && (
-              <>
-                {restRooms.length === 0 ? (
-                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                    此分類暫無資料
-                  </div>
-                ) : (
-                  restRooms.map((station) => (
-                    <InfoCard
-                      key={station.id}
-                      type={station.facility_type}
-                      name={station.name}
-                      address={station.location}
-                      contact={station.phone}
-                      hours={station.opening_hours || ""}
-                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
-                        station.location
-                      )}`}
-                      fullData={station}
-                    />
-                  ))
-                )}
-              </>
-            )}
-
-            {!loading && !error && selectedCategory === "medical_stations" && (
-              <>
-                {medicalStations.length === 0 ? (
-                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                    此分類暫無資料
-                  </div>
-                ) : (
-                  medicalStations.map((station) => (
-                    <InfoCard
-                      key={station.id}
-                      type={station.station_type}
-                      name={station.name}
-                      address={station.location}
-                      contact={station.phone}
-                      hours={station.operating_hours || ""}
-                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
-                        station.location
-                      )}`}
-                      fullData={station}
-                    />
-                  ))
-                )}
-              </>
-            )}
-
-            {!loading && !error && selectedCategory === "accommodations" && (
-              <>
-                {accommodations.length === 0 ? (
-                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
-                    此分類暫無資料
-                  </div>
-                ) : (
-                  accommodations.map((station) => (
-                    <InfoCard
-                      key={station.id}
-                      name={station.name}
-                      address={station.location}
-                      contact={station.contact_info}
-                      hours={station.available_period || ""}
-                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
-                        station.location
-                      )}`}
-                      fullData={station}
-                    />
-                  ))
-                )}
-              </>
-            )}
-          </div>
-        )}
-      </div>
-      <br />
-    </div>
-  );
-}
