+++ src/features/SiteMap/index.tsx
@@ -1,119 +1,413 @@
 "use client";
 
-import { useState } from "react";
-import MapShow from "@/features/SiteMap/MapShow";
+import React, { useState } from "react";
+import Button from "@/components/Button";
+import {
+  getWaterRefillStations,
+  getShowerStations,
+  getRestrooms,
+  getMedicalStations,
+  getAccommodations,
+  type WaterRefillStations,
+  type ShowerStations,
+  type RestRooms,
+  type MedicalStation,
+  type Accommodations,
+} from "@/lib/api";
+import InfoCard from "@/components/InfoCard";
+
+type LocationCategory =
+  | "all"
+  | "water_refill_stations"
+  | "shower_stations"
+  | "restrooms"
+  | "medical_stations"
+  | "accommodations";
+type ShowMode = "mapShow" | "listShow";
+
+const CATEGORIES = [
+  {
+    key: "all",
+    name: "全部",
+  },
+  {
+    key: "water_refill_stations",
+    name: "加水站",
+  },
+  {
+    key: "shower_stations",
+    name: "洗澡點",
+  },
+  {
+    key: "restrooms",
+    name: "廁所",
+  },
+  {
+    key: "medical_stations",
+    name: "醫療站",
+  },
+  {
+    key: "accommodations",
+    name: "住宿",
+  },
+];
+
+const MAP_URL = "https://guangfu250923-map.pttapp.cc/map.html";
+const MAP_HEIGHT = 422;
 
 export default function SiteMap() {
-  const [showWarningModal, setShowWarningModal] = useState(false);
+  const [showMode, setShowMode] = useState<ShowMode>("mapShow");
+  const [selectedCategory, setSelectedCategory] =
+    useState<LocationCategory>("all");
+  const [waterRefillStations, setWaterRefillStations] = useState<
+    WaterRefillStations[]
+  >([]);
+  const [showerStations, setShowerStations] = useState<ShowerStations[]>([]);
+  const [restRooms, setRestRooms] = useState<RestRooms[]>([]);
+  const [medicalStations, setMedicalStations] = useState<MedicalStation[]>([]);
+  const [accommodations, setAccommodations] = useState<Accommodations[]>([]);
+  const [allData, setAllData] = useState<
+    (
+      | WaterRefillStations
+      | ShowerStations
+      | RestRooms
+      | MedicalStation
+      | Accommodations
+    )[]
+  >([]);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState<string | null>(null);
+
+  async function fetchWaterRefillStations() {
+    try {
+      setLoading(true);
+      setError(null);
+      const response = await getWaterRefillStations(50, 0);
+      setWaterRefillStations(response.member);
+    } catch (err) {
+      setError(err instanceof Error ? err.message : "載入失敗");
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  async function fetchShowerStations() {
+    try {
+      setLoading(true);
+      setError(null);
+      const response = await getShowerStations(50, 0);
+      setShowerStations(response.member);
+    } catch (err) {
+      setError(err instanceof Error ? err.message : "載入失敗");
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  async function fetchRestRooms() {
+    try {
+      setLoading(true);
+      setError(null);
+      const response = await getRestrooms(50, 0);
+      setRestRooms(response.member);
+    } catch (err) {
+      setError(err instanceof Error ? err.message : "載入失敗");
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  async function fetchMedicalStations() {
+    try {
+      setLoading(true);
+      setError(null);
+      const response = await getMedicalStations(50, 0);
+      setMedicalStations(response.member);
+    } catch (err) {
+      setError(err instanceof Error ? err.message : "載入失敗");
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  async function fetchAccommodations() {
+    try {
+      setLoading(true);
+      setError(null);
+      const response = await getAccommodations(50, 0);
+      setAccommodations(response.member);
+    } catch (err) {
+      setError(err instanceof Error ? err.message : "載入失敗");
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  async function fetchAll() {
+    try {
+      setLoading(true);
+      setError(null);
+      const [
+        responseWaterRefillStations,
+        responseShowerStations,
+        responseRestrooms,
+        responseMedicalStations,
+        responseAccommodations,
+      ] = await Promise.all([
+        getWaterRefillStations(50, 0),
+        getShowerStations(50, 0),
+        getRestrooms(50, 0),
+        getMedicalStations(50, 0),
+        getAccommodations(50, 0),
+      ]);
+      setWaterRefillStations(responseWaterRefillStations.member);
+      setShowerStations(responseShowerStations.member);
+      setRestRooms(responseRestrooms.member);
+      setMedicalStations(responseMedicalStations.member);
+      setAccommodations(responseAccommodations.member);
+
+      const combined = [
+        ...responseWaterRefillStations.member,
+        ...responseShowerStations.member,
+        ...responseRestrooms.member,
+        ...responseMedicalStations.member,
+        ...responseAccommodations.member,
+      ];
+      combined.sort((a, b) => a.created_at - b.created_at);
+      setAllData(combined);
+    } catch (err) {
+      setError(err instanceof Error ? err.message : "載入失敗");
+    } finally {
+      setLoading(false);
+    }
+  }
+
+  const handleCategoryClick = async (categoryKey: LocationCategory) => {
+    setSelectedCategory(categoryKey);
+    if (categoryKey === "all") {
+      await fetchAll();
+    } else if (categoryKey === "water_refill_stations") {
+      await fetchWaterRefillStations();
+    } else if (categoryKey === "shower_stations") {
+      await fetchShowerStations();
+    } else if (categoryKey === "restrooms") {
+      await fetchRestRooms();
+    } else if (categoryKey === "medical_stations") {
+      await fetchMedicalStations();
+    } else if (categoryKey === "accommodations") {
+      await fetchAccommodations();
+    }
+  };
+
+  const handleModeChange = (event: React.ChangeEvent<HTMLSelectElement>) => {
+    setShowMode(event.target.value as ShowMode);
+    fetchAll();
+  };
 
   return (
     <div>
-      {/* 警示條 */}
-      <div
-        className="bg-[#F23555] p-4 overflow-hidden cursor-pointer"
-        onClick={() => setShowWarningModal(true)}
-      >
-        <div className="flex items-start gap-3">
-          <div className="flex-1 overflow-hidden">
-            <span className="inline-block text-[#FFFFFF] font-medium">
-              現場隨時有溢流風險，點此詳讀避難守則，警報響起請往高處避難
-            </span>
-          </div>
-        </div>
-      </div>
-
-      {/* 警告視窗 */}
-      {showWarningModal && (
-        <div className="fixed inset-0 z-50 flex items-center justify-center p-4">
-          <div
-            className="fixed inset-0 bg-black/20 backdrop-blur-[4px]"
-            onClick={() => setShowWarningModal(false)}
-          />
-          <div className="relative bg-[var(--text-black)] rounded-[20px] max-w-md w-full p-6 shadow-2xl flex flex-col gap-6">
-            <header className="h-[25px] flex items-center">
-              {/* 關閉按鈕 */}
-              <button
-                onClick={() => setShowWarningModal(false)}
-                className="text-white hover:text-gray-300 text-2xl w-6 h-6 flex items-center justify-center cursor-pointer"
-                aria-label="關閉視窗"
+      <div className="flex my-3">
+        <select
+          className="bg-gray-100 p-3"
+          value={showMode}
+          onChange={handleModeChange}
+        >
+          <option value="mapShow">地圖顯示</option>
+          <option value="listShow">列表顯示</option>
+        </select>
+        {showMode === "listShow" && (
+          <div className="ml-4 flex overflow-y-scroll [scrollbar-width:none]">
+            {CATEGORIES.map((category) => (
+              <Button
+                className="ml-2 border-gray-100"
+                key={category.key}
+                onClick={() =>
+                  handleCategoryClick(category.key as LocationCategory)
+                }
+                active={selectedCategory === category.key}
               >
-                ✕
-              </button>
-
-              {/* 標題 */}
-              <h2 className="flex-1 h-[25px] text-[20px] font-medium text-white text-center leading-[25px]">
-                避難守則
-              </h2>
-            </header>
-
-            {/* 內容 */}
-            <div className="text-white">
-              <p className="text-2xl font-medium text-center text-white">
-                堰塞湖隨時有溢流風險
-              </p>
-              <p className="text-[28px] leading-[35px] font-bold text-center text-[#F23555]">
-                若警報響起，請儘速往
-                <br />
-                高處或高樓層避難
-              </p>
-            </div>
-
-            {/* 警報說明 */}
-            <div className="flex flex-col gap-2">
-              <div className="flex items-center gap-5 bg-[#343434] rounded-[16px] p-4">
-                <div className="h-12 w-12 flex items-center justify-center">
-                  <svg
-                    width="36"
-                    height="36"
-                    viewBox="0 0 36 36"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      d="M22 35.9502V31.8502C25 30.9835 27.4167 29.3169 29.25 26.8502C31.0833 24.3835 32 21.5835 32 18.4502C32 15.3169 31.0833 12.5169 29.25 10.0502C27.4167 7.58353 25 5.91686 22 5.0502V0.950195C26.1333 1.88353 29.5 3.9752 32.1 7.2252C34.7 10.4752 36 14.2169 36 18.4502C36 22.6835 34.7 26.4252 32.1 29.6752C29.5 32.9252 26.1333 35.0169 22 35.9502ZM0 24.5002V12.5002H8L18 2.5002V34.5002L8 24.5002H0ZM22 26.5002V10.4002C23.5667 11.1335 24.7917 12.2335 25.675 13.7002C26.5583 15.1669 27 16.7669 27 18.5002C27 20.2002 26.5583 21.7752 25.675 23.2252C24.7917 24.6752 23.5667 25.7669 22 26.5002ZM14 12.2002L9.7 16.5002H4V20.5002H9.7L14 24.8002V12.2002Z"
-                      fill="white"
-                    />
-                  </svg>
-                </div>
-                <div className="flex-1 text-white">
-                  <p className="font-bold h-[25px]">海嘯警報（無語音）</p>
-                  <p className="text-base leading-5">鳴5秒、停5秒</p>
-                  <p className="text-base leading-5">重複9遍，共85秒</p>
-                </div>
-              </div>
-
-              <div className="flex items-center gap-5 bg-[#343434] rounded-[16px] p-4">
-                <div className="h-12 w-12 flex items-center justify-center">
-                  <svg
-                    width="36"
-                    height="36"
-                    viewBox="0 0 36 36"
-                    fill="none"
-                    xmlns="http://www.w3.org/2000/svg"
-                  >
-                    <path
-                      d="M22 35.9502V31.8502C25 30.9835 27.4167 29.3169 29.25 26.8502C31.0833 24.3835 32 21.5835 32 18.4502C32 15.3169 31.0833 12.5169 29.25 10.0502C27.4167 7.58353 25 5.91686 22 5.0502V0.950195C26.1333 1.88353 29.5 3.9752 32.1 7.2252C34.7 10.4752 36 14.2169 36 18.4502C36 22.6835 34.7 26.4252 32.1 29.6752C29.5 32.9252 26.1333 35.0169 22 35.9502ZM0 24.5002V12.5002H8L18 2.5002V34.5002L8 24.5002H0ZM22 26.5002V10.4002C23.5667 11.1335 24.7917 12.2335 25.675 13.7002C26.5583 15.1669 27 16.7669 27 18.5002C27 20.2002 26.5583 21.7752 25.675 23.2252C24.7917 24.6752 23.5667 25.7669 22 26.5002ZM14 12.2002L9.7 16.5002H4V20.5002H9.7L14 24.8002V12.2002Z"
-                      fill="white"
-                    />
-                  </svg>
-                </div>
-                <div className="flex-1 text-white">
-                  <p className="font-bold h-[25px]">解除警報</p>
-                  <p className="text-base leading-5">一長聲90秒</p>
-                </div>
-              </div>
-            </div>
-
-            <div className="text-[16px] leading-[20px] text-center text-white">
-              如未聽到警報，請勿誤信錯誤訊息！
-              <br />
-              並請大家協助轉傳正確資訊
-            </div>
+                {category.name}
+              </Button>
+            ))}
           </div>
-        </div>
-      )}
+        )}
+      </div>
+      <div>
+        {showMode === "mapShow" && (
+          <iframe
+            src={MAP_URL}
+            title="地圖顯示"
+            width="100%"
+            height={MAP_HEIGHT}
+          />
+        )}
+        {showMode === "listShow" && (
+          <div className="space-y-4">
+            {loading && (
+              <div className="text-center py-8 text-gray-500 dark:text-gray-400">
+                載入中...
+              </div>
+            )}
 
-      <MapShow />
+            {error && (
+              <div className="text-center py-8 text-red-500 dark:text-red-400">
+                錯誤: {error}
+              </div>
+            )}
+
+            {!loading && !error && selectedCategory === "all" && (
+              <>
+                {allData.length === 0 ? (
+                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
+                    此分類暫無資料
+                  </div>
+                ) : (
+                  allData.map((station) => (
+                    <InfoCard
+                      key={station.id}
+                      name={station.name}
+                      address={station.location}
+                      contact={station.phone}
+                      hours={station.opening_hours || ""}
+                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
+                        station.location
+                      )}`}
+                      fullData={station}
+                    />
+                  ))
+                )}
+              </>
+            )}
+
+            {!loading &&
+              !error &&
+              selectedCategory === "water_refill_stations" && (
+                <>
+                  {waterRefillStations.length === 0 ? (
+                    <div className="text-center py-8 text-gray-500 dark:text-gray-400">
+                      此分類暫無資料
+                    </div>
+                  ) : (
+                    waterRefillStations.map((station) => (
+                      <InfoCard
+                        key={station.id}
+                        name={station.name}
+                        type={station.water_type}
+                        address={station.location}
+                        contact={station.phone}
+                        hours={station.opening_hours || ""}
+                        mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
+                          station.location
+                        )}`}
+                        fullData={station}
+                      />
+                    ))
+                  )}
+                </>
+              )}
+
+            {!loading && !error && selectedCategory === "shower_stations" && (
+              <>
+                {showerStations.length === 0 ? (
+                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
+                    此分類暫無資料
+                  </div>
+                ) : (
+                  showerStations.map((station) => (
+                    <InfoCard
+                      key={station.id}
+                      type={station.facility_type}
+                      name={station.name}
+                      address={station.location}
+                      contact={station.phone}
+                      hours={station.time_slots || ""}
+                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
+                        station.location
+                      )}`}
+                      fullData={station}
+                    />
+                  ))
+                )}
+              </>
+            )}
+
+            {!loading && !error && selectedCategory === "restrooms" && (
+              <>
+                {restRooms.length === 0 ? (
+                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
+                    此分類暫無資料
+                  </div>
+                ) : (
+                  restRooms.map((station) => (
+                    <InfoCard
+                      key={station.id}
+                      type={station.facility_type}
+                      name={station.name}
+                      address={station.location}
+                      contact={station.phone}
+                      hours={station.opening_hours || ""}
+                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
+                        station.location
+                      )}`}
+                      fullData={station}
+                    />
+                  ))
+                )}
+              </>
+            )}
+
+            {!loading && !error && selectedCategory === "medical_stations" && (
+              <>
+                {medicalStations.length === 0 ? (
+                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
+                    此分類暫無資料
+                  </div>
+                ) : (
+                  medicalStations.map((station) => (
+                    <InfoCard
+                      key={station.id}
+                      type={station.station_type}
+                      name={station.name}
+                      address={station.location}
+                      contact={station.phone}
+                      hours={station.operating_hours || ""}
+                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
+                        station.location
+                      )}`}
+                      fullData={station}
+                    />
+                  ))
+                )}
+              </>
+            )}
+
+            {!loading && !error && selectedCategory === "accommodations" && (
+              <>
+                {accommodations.length === 0 ? (
+                  <div className="text-center py-8 text-gray-500 dark:text-gray-400">
+                    此分類暫無資料
+                  </div>
+                ) : (
+                  accommodations.map((station) => (
+                    <InfoCard
+                      key={station.id}
+                      name={station.name}
+                      address={station.location}
+                      contact={station.contact_info}
+                      hours={station.available_period || ""}
+                      mapUrl={`https://www.google.com/maps/search/?api=1&query=${encodeURIComponent(
+                        station.location
+                      )}`}
+                      fullData={station}
+                    />
+                  ))
+                )}
+              </>
+            )}
+          </div>
+        )}
+      </div>
+      <br />
     </div>
   );
 }
